% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_help.R
\name{rmd_help}
\alias{rmd_help}
\title{Browse a help file as an Rmd}
\usage{
rmd_help(topic, package = NULL)
}
\arguments{
\item{topic}{bare symbol to search for help on. pkg::func syntax is supported and if used \code{package} is ignored.}

\item{package}{package name to resolve symbol in}
}
\value{
nothing. Opens help as side effect.
}
\description{
A drop-in replacement for \code{help()} that opens the help file as Rmd.
}
\details{
You're better off binding this to a key or using the RStudio addin. See
README.
}
\examples{
\dontrun{
rmd_help(help)
}
}
