% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{rs_rmd_help}
\alias{rs_rmd_help}
\title{RMarkdown help() on cursor word or selection}
\usage{
rs_rmd_help()
}
\value{
nothing. Opens document as side effect
}
\description{
Analagous to the default \code{F1} shortcut in RStudio, except it opens the help
for the thing the cursor is on, or is selected, in Rmd format.
}
\details{
Bind the 'RMarkdown help() on object addin to a keyboard shortcut to use
this.
}
