% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_help.R
\name{rmd_help}
\alias{rmd_help}
\title{Browse a help file as an Rmd}
\usage{
rmd_help(topic)
}
\arguments{
\item{topic}{bare symbol to search for help on.}
}
\value{
nothing
}
\description{
A drop-in replacement for \code{help()}.
}
\details{
FUNCTION_DESCRIPTION
}
\examples{
rmd_help(help)
}
