% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_help.R
\name{rmd_help}
\alias{rmd_help}
\title{Browse a help file as an Rmd}
\usage{
rmd_help(topic)
}
\arguments{
\item{topic}{bare symbol to search for help on.}
}
\value{
nothing. Opens help as side effect.
}
\description{
A drop-in replacement for \code{help()} that opens the help file as Rmd.
}
\details{
You're better off binding this to a key. See README.
}
\examples{
rmd_help(help)
}
